refactor: shuffleAll
